Moved the code around to create the correct structure for the remote step
